Create README.md
MITM Bus Pass Management System - This is a project which is done for generating and maintaining the bus passes for the students of college MITM. The students of this college can also do some enquiry regarding their bus passes or regarding their travelling if required using this application.
It became easier for the admin(The one who handles ths application) to handle the information of all the students with bus passes as the admin can maintain it online and access theses information easily when required. If this was to be done manually by admin it would have been difficult and tome consuming.
So this is an interesting project.